Revert "PROYECTO FINALIZADO"
This reverts commit 9db7b6dd611a3958474ea9391a47ca3a27e6ab1c.

+++ JAVA/VanguardBank2023/src/java/interfaces/CuentaCRUD.java
@@ -11,7 +11,7 @@ import model.Cuenta;
 public interface CuentaCRUD {
     public List listarCuentas();
     public boolean agregarCuenta(Cuenta nuevaCuenta);
-    public Cuenta buscarCuenta(int numCuenta);
+    public Cuenta buscarCuenta(long numCuenta);
     public boolean editarCuenta(Cuenta cuenta);
     public void eliminarCuenta(long numCuenta);
 }
